Add vaos-vbos.cpp methods
In this file, we completed the code for creating the descriptors for VBO attributes
and indices (DescrVBOAtrib and DescrVBOInds classes), specifically, we completed the
crearVBO method in each of these two classes.

We also completed the crearVAO method of the DescrVAO class, writing the code that
follows these steps.

Finally, we implemented the code for the draw method of the DescrVAO class. This method
is responsible for rendering the VAO. The first time it is executed for a specific descriptor,
the VAO will be created in the GPU before rendering. At the end of the method, the VAO
is deactivated to avoid affecting future operations.